eth_validators/requirements.txt
click
PyYAML
requests
tabulate
